Update Stacked2dCore's batch_norm
To achieve a variance of 1 after batch normalisation the scale parameter must not be learnable. Though pytorch doesn't allow learning the batch_norm bias without learning the scale factor, batch_norm must be allowed being non-learnable. The bias must then already be learned in the conv layer. The proposed changes enable just that, with it also being default (as Santiago and Max handled it).